refactor(cell): split ticks.rs (1010 lines) into ticks/ directory module
Extract along natural lifecycle seams:

- mod.rs: run_aoi_tick, reload_completion_tick, pending_attack_tick,
  pending_reload_tick, regen_tick + their tests
- holster.rs: HOLSTER_ANIMATION_DURATION constant, holster_timer_tick,
  pending_slot_swap_tick + holster tests
- npc_movement.rs: npc_movement_tick + movement tests

All 269 original comment lines preserved exactly.
