+++ crates/services/src/cell/service/ticks/npc_movement.rs
@@ -1,0 +1,187 @@
+use super::super::super::space_manager::SpaceManager;
+
+/// NPC movement along nav paths — runs every AoI tick (100ms) for smooth pathing.
+///
+/// For each NPC with a non-empty `nav_path`, move it toward the next waypoint
+/// by `move_speed` units. When it reaches (or overshoots) a waypoint, consume
+/// it and continue to the next. Position updates propagate to witnesses via
+/// the AoI tick's `EntityMoved` messages.
+pub(in crate::cell::service) fn npc_movement_tick(space_mgr: &mut SpaceManager) {
+    // Collect NPCs that have active paths
+    let moving_npcs: Vec<u32> = space_mgr
+        .all_npc_entity_ids()
+        .iter()
+        .filter(|&&eid| {
+            space_mgr
+                .get_entity(eid)
+                .is_some_and(|e| !e.nav_path.is_empty())
+        })
+        .copied()
+        .collect();
+
+    for npc_id in moving_npcs {
+        // Read the next waypoint, move_speed, and remaining path length
+        let (next_wp, move_speed, cur_pos, path_len) = {
+            let npc = match space_mgr.get_entity(npc_id) {
+                Some(e) if !e.nav_path.is_empty() => e,
+                _ => continue,
+            };
+            let next_wp = match npc.nav_path.front() {
+                Some(wp) => *wp,
+                None => continue,
+            };
+            (next_wp, npc.move_speed, npc.position, npc.nav_path.len())
+        };
+
+        let dx = next_wp.x - cur_pos.x;
+        let dy = next_wp.y - cur_pos.y;
+        let dz = next_wp.z - cur_pos.z;
+        let dist = (dx * dx + dy * dy + dz * dz).sqrt();
+
+        // Speed in world units per second (tick is 100ms = 0.1s)
+        let speed_per_sec = move_speed * 10.0;
+
+        if dist <= move_speed {
+            // Reached (or overshot) the waypoint — snap to it and consume
+            // Waypoint Y comes from Detour's findStraightPath (already on navmesh surface)
+            let snap_y = next_wp.y;
+
+            // Peek at the NEXT waypoint (index 1) to compute velocity toward it
+            let next_next_wp = if path_len > 1 {
+                space_mgr
+                    .get_entity(npc_id)
+                    .and_then(|e| e.nav_path.get(1).copied())
+            } else {
+                None
+            };
+
+            let (velocity, yaw) = if let Some(nn) = next_next_wp {
+                // Still more waypoints — compute velocity toward the next one
+                let ndx = nn.x - next_wp.x;
+                let ndz = nn.z - next_wp.z;
+                let ndy = nn.y - next_wp.y;
+                let nd = (ndx * ndx + ndy * ndy + ndz * ndz).sqrt();
+                if nd > 0.001 {
+                    (
+                        [
+                            ndx / nd * speed_per_sec,
+                            ndy / nd * speed_per_sec,
+                            ndz / nd * speed_per_sec,
+                        ],
+                        ndx.atan2(ndz),
+                    )
+                } else {
+                    ([0.0; 3], 0.0)
+                }
+            } else {
+                // Last waypoint — stopping, keep current facing
+                ([0.0; 3], dx.atan2(dz))
+            };
+
+            space_mgr.update_entity_position(
+                npc_id,
+                [next_wp.x, snap_y, next_wp.z],
+                [0, 0, 0],
+                velocity,
+            );
+            if let Some(npc) = space_mgr.get_entity_mut(npc_id) {
+                npc.nav_path.pop_front();
+                npc.direction = cimmeria_common::Vector3::new(0.0, yaw, 0.0);
+            }
+        } else {
+            // Move toward waypoint by move_speed units
+            let t = move_speed / dist;
+            let new_x = cur_pos.x + dx * t;
+            let new_z = cur_pos.z + dz * t;
+
+            // Linearly interpolate Y between current position and waypoint.
+            // Waypoints from Detour's findStraightPath are on the navmesh surface,
+            // so linear interpolation between them stays close to the floor.
+            let new_y = cur_pos.y + dy * t;
+
+            // Face the direction of movement (yaw = atan2(dx, dz) in radians)
+            // Direction is [pitch, yaw, roll] — only yaw matters for facing
+            let yaw = dx.atan2(dz);
+
+            // Velocity = direction * speed_per_sec
+            let velocity = [
+                dx / dist * speed_per_sec,
+                dy / dist * speed_per_sec,
+                dz / dist * speed_per_sec,
+            ];
+
+            if (npc_id % 10000) < 5 {
+                // log a few NPCs
+                tracing::debug!(
+                    npc_id,
+                    cur = format_args!("({:.1},{:.1},{:.1})", cur_pos.x, cur_pos.y, cur_pos.z),
+                    new = format_args!("({:.1},{:.1},{:.1})", new_x, new_y, new_z),
+                    wp = format_args!("({:.1},{:.1},{:.1})", next_wp.x, next_wp.y, next_wp.z),
+                    "NPC movement step"
+                );
+            }
+
+            space_mgr.update_entity_position(npc_id, [new_x, new_y, new_z], [0, 0, 0], velocity);
+            // Set yaw directly as radians (pack_angle reads direction.y)
+            if let Some(npc) = space_mgr.get_entity_mut(npc_id) {
+                npc.direction = cimmeria_common::Vector3::new(0.0, yaw, 0.0);
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::cell::space_manager::SpaceManager;
+
+    #[test]
+    fn npc_movement_tick_advances_along_nav_path() {
+        let mut mgr = SpaceManager::new(1);
+        let xml = r#"<?xml version="1.0"?><Spaces><Space WorldName="Castle" Instanced="false" MinX="-800" MaxX="800" MinY="-800" MaxY="800" /></Spaces>"#;
+        mgr.parse_spaces_xml(xml).unwrap();
+        mgr.create_startup_spaces(
+            r#"<?xml version="1.0"?><Spaces><Space WorldName="Castle" /></Spaces>"#,
+        )
+        .unwrap();
+        mgr.create_entity(200, "Castle", [0.0, 0.0, 0.0], [0.0; 3])
+            .unwrap();
+        if let Some(npc) = mgr.get_entity_mut(200) {
+            npc.is_player = false;
+            npc.class_id = 0x04;
+            npc.move_speed = 5.0;
+            npc.nav_path
+                .push_back(cimmeria_common::Vector3::new(10.0, 0.0, 0.0));
+        }
+
+        npc_movement_tick(&mut mgr);
+
+        let npc = mgr.get_entity(200).unwrap();
+        assert_eq!(npc.position.x, 5.0);
+        assert_eq!(npc.position.y, 0.0);
+        assert_eq!(npc.position.z, 0.0);
+        assert_eq!(npc.nav_path.len(), 1);
+    }
+
+    #[test]
+    fn npc_movement_tick_does_not_panic_on_empty_path() {
+        let mut mgr = SpaceManager::new(1);
+        let xml = r#"<?xml version="1.0"?><Spaces><Space WorldName="Castle" Instanced="false" MinX="-800" MaxX="800" MinY="-800" MaxY="800" /></Spaces>"#;
+        mgr.parse_spaces_xml(xml).unwrap();
+        mgr.create_startup_spaces(
+            r#"<?xml version="1.0"?><Spaces><Space WorldName="Castle" /></Spaces>"#,
+        )
+        .unwrap();
+        mgr.create_entity(200, "Castle", [0.0, 0.0, 0.0], [0.0; 3])
+            .unwrap();
+        if let Some(npc) = mgr.get_entity_mut(200) {
+            npc.is_player = false;
+            npc.class_id = 0x04;
+            npc.nav_path.clear();
+        }
+        // Must not panic.
+        npc_movement_tick(&mut mgr);
+        let npc = mgr.get_entity(200).unwrap();
+        assert_eq!(npc.position.x, 0.0, "stationary NPC must not move");
+    }
+}
